requirements.txt
streamlit
openai
instructor
pydantic
python-dotenv
ratelimit
google-generativeai